// pro/src/apiClient/adage/models/CollectiveAdditionalFeeType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * An enumeration.
 */
export enum CollectiveAdditionalFeeType {
  ACCOMMODATION = 'ACCOMMODATION',
  TRAVEL = 'TRAVEL',
  MEAL = 'MEAL',
  CONSUMABLE_ITEMS = 'CONSUMABLE_ITEMS',
  COPYRIGHT = 'COPYRIGHT',
  APPLICATION_FEE = 'APPLICATION_FEE',
  OTHER = 'OTHER',
}